+++ src/js/src/utils/errorNotif.js
@@ -36,7 +36,6 @@ export async function useErrorNotification(error) {
     hasIcon: true,
     position: 'is-top',
     indefinite: true,
-    ariaCloseLabel: 'Dismiss error',
     message: message,
   });
 }