+++ src/js/src/views/Experiments.vue
@@ -199,7 +199,7 @@
           default-sort="name">
           <template #empty>
             <section class="section">
-              <div class="content has-text-white has-text-centered">
+              <div class="content has-text-centered">
                 Your search turned up empty!
               </div>
             </section>