+++ src/js/src/views/Users.vue
@@ -150,11 +150,7 @@
                 readonly></b-input>
             </b-field>
 
-            <b-button
-              size="is-small"
-              icon-left="copy"
-              aria-label="Copy token"
-              @click="copy">
+            <b-button size="is-small" icon-left="copy" @click="copy">
             </b-button>
 
             <b-field label="Expires">
@@ -200,10 +196,7 @@
     <b-field grouped position="is-right">
       <p v-if="roleAllowed('users', 'create')" class="control">
         <b-tooltip label="create a new user" type="is-light is-left">
-          <button
-            class="button is-light"
-            aria-label="Create a new user"
-            @click="isCreateActive = true">
+          <button class="button is-light" @click="isCreateActive = true">
             <b-icon icon="plus"></b-icon>
           </button>
         </b-tooltip>
@@ -250,7 +243,6 @@
             multilined>
             <button
               class="button is-light is-small"
-              :aria-label="`Create token for ${props.row.username}`"
               @click="newToken(props.row.username)">
               <b-icon icon="key"></b-icon>
             </button>
@@ -264,7 +256,6 @@
             <button
               v-if="roleAllowed('users', 'delete', props.row.username)"
               class="button is-light is-small"
-              :aria-label="`Delete user ${props.row.username}`"
               @click="deleteUser(props.row.username)">
               <b-icon icon="trash"></b-icon>
             </button>
@@ -278,7 +269,6 @@
             <button
               v-if="roleAllowed('users', 'patch', props.row.username)"
               class="button is-light is-small"
-              :aria-label="`Edit user ${props.row.username}`"
               @click="editUser(props.row.username)">
               <b-icon icon="pencil"></b-icon>
             </button>